feat(clickhouse): 调整K线和交易数据表字段类型为整数以提升性能 同时在数据写入时通过 roundEx 对价格和成交量进行四舍五入处理，并转换为整数存储。 读取时通过乘以 0.001 恢复原始数值，确保精度不丢失。 修复了导入数据时对成交量比较的逻辑错误，增强数据一致性校验。

+++ hikyuu/data/clickhouse_upgrade/createdb.sql
@@ -883,12 +883,12 @@ CREATE table if not exists hku_data.day_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -899,12 +899,12 @@ CREATE table if not exists hku_data.week_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -914,12 +914,12 @@ CREATE table if not exists hku_data.month_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -929,12 +929,12 @@ CREATE table if not exists hku_data.quarter_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -944,12 +944,12 @@ CREATE table if not exists hku_data.halfyear_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -959,12 +959,12 @@ CREATE table if not exists hku_data.year_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -974,12 +974,12 @@ CREATE table if not exists hku_data.min_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date)-toYear(date)%10)
@@ -989,12 +989,12 @@ CREATE table if not exists hku_data.min5_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date)-toYear(date)%10)
@@ -1004,12 +1004,12 @@ CREATE table if not exists hku_data.min15_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date)-toYear(date)%10)
@@ -1019,12 +1019,12 @@ CREATE table if not exists hku_data.min30_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date)-toYear(date)%10)
@@ -1034,12 +1034,12 @@ CREATE table if not exists hku_data.min60_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -1049,12 +1049,12 @@ CREATE table if not exists hku_data.hour2_k (
     `market` String,
     `code` String,
     `date` DateTime,
-    `open` DOUBLE,
-    `high` DOUBLE,
-    `low` DOUBLE,
-    `close` DOUBLE,
-    `amount` DOUBLE,
-    `volume` DOUBLE
+    `open` UInt32,
+    `high` UInt32,
+    `low` UInt32,
+    `close` UInt32,
+    `amount` UInt64,
+    `volume` UInt64
 ) 
 ENGINE = MergeTree()
 PARTITION BY market
@@ -1064,8 +1064,8 @@ CREATE table if not exists hku_data.timeline (
     `market` String,
     `code` String,
     `date` DateTime,
-    `price` DOUBLE,
-    `vol` DOUBLE
+    `price` UInt32,
+    `vol` UInt32
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date)-toYear(date)%10)
@@ -1075,9 +1075,9 @@ CREATE table if not exists hku_data.transdata (
     `market` String,
     `code` String,
     `date` DateTime,
-    `price` DOUBLE,
-    `vol` DOUBLE,
-    `buyorsell` int
+    `price` UInt32,
+    `vol` UInt32,
+    `buyorsell` UInt8
 ) 
 ENGINE = MergeTree()
 PARTITION BY (market, toYear(date))
